[refactor] #193 보관 목표 관련 파일 패키징
