+++ app/src/main/AndroidManifest.xml
@@ -65,7 +65,7 @@
             android:screenOrientation="portrait"
             tools:ignore="LockedOrientationActivity" />
         <activity
-            android:name=".presentation.my.ArchivedGoalActivity"
+            android:name=".presentation.my.archive.ArchivedGoalActivity"
             android:exported="false"
             android:screenOrientation="portrait"
             tools:ignore="LockedOrientationActivity" />
