+++ app/src/main/java/org/keepgoeat/presentation/my/archive/ArchivedGoalActivity.kt
@@ -1,4 +1,4 @@
-package org.keepgoeat.presentation.my
+package org.keepgoeat.presentation.my.archive
 
 import android.content.Intent
 import android.os.Bundle
@@ -15,6 +15,9 @@ import kotlinx.coroutines.flow.onEach
 import org.keepgoeat.R
 import org.keepgoeat.databinding.ActivityArchivedGoalBinding
 import org.keepgoeat.presentation.home.HomeActivity
+import org.keepgoeat.presentation.my.GoalDeleteDialogFragment
+import org.keepgoeat.presentation.my.MyActivity
+import org.keepgoeat.presentation.my.MyViewModel
 import org.keepgoeat.presentation.type.EatingType
 import org.keepgoeat.presentation.type.SortType
 import org.keepgoeat.util.UiState
